prototype: split test modules
