+++ test/lsm-prototypes-tests.hs
@@ -3,10 +3,14 @@ module Main (main) where
 import           Test.Tasty
 
 import qualified FormatPage
+import qualified ScheduledMergesTest
 import qualified ScheduledMergesTestQLS
 
 main :: IO ()
 main = defaultMain $ testGroup "prototype" [
-      ScheduledMergesTestQLS.tests,
+      testGroup "ScheduledMerges" [
+        ScheduledMergesTest.tests,
+        ScheduledMergesTestQLS.tests
+      ],
       FormatPage.tests
     ]
